More C++11 correctness and nicer test filename
