+++ test/test_cpp_double_float_constructors.cpp
@@ -6,7 +6,7 @@
 //  Software License, Version 1.0. (See accompanying file
 //  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
-// Basic compilation, constructors, and operatory tests for cpp_double_float<>
+// Constructor tests for cpp_double_float<>
 
 #include <boost/multiprecision/cpp_double_float.hpp>
 #include <boost/multiprecision/cpp_bin_float.hpp>
@@ -30,18 +30,43 @@ FloatingPointType uniform_real()
 
 template <typename NumericType,
           typename std::enable_if<std::is_integral<NumericType>::value, bool>::type = true>
-NumericType uniform_integral_number() {
+NumericType uniform_integral_number()
+{
    NumericType out = 0;
-   
+
    for (int i = 0; i < sizeof(NumericType); ++i)
       out = (out << 8) + static_cast<NumericType>(std::round(256.0 * uniform_real<float>()));
 
    return out;
 }
 
-template <typename FloatingPointType, typename NumericType>
-int test_constructor()
+
+template <typename NumericType,
+          typename std::enable_if<std::is_integral<NumericType>::value && !std::is_floating_point<NumericType>::value, bool>::type = true>
+NumericType get_rand()
 {
+   return uniform_integral_number<NumericType>();
+}
+
+template <typename FloatingPointType,
+          typename std::enable_if<std::is_floating_point<FloatingPointType>::value, bool>::type = true>
+FloatingPointType get_rand()
+{
+   return uniform_real<FloatingPointType>();
+}
+
+template <typename FloatingPointType>
+boost::multiprecision::backends::cpp_double_float<typename FloatingPointType::float_type> get_rand()
+{
+   using float_type = typename FloatingPointType::float_type;
+   return boost::multiprecision::backends::cpp_double_float<float_type>(uniform_real<float_type>()) * boost::multiprecision::backends::cpp_double_float<float_type>(uniform_real<float_type>());
+}
+
+template <typename FloatingPointType, typename NumericType, typename std::enable_if<std::is_integral<NumericType>::value>::type const* = nullptr>
+bool test_constructor()
+{
+   bool result_is_ok = true;
+
    constexpr int Trials = 50000;
    std::cout << "Testing constructor for ";
    std::cout.width(30);
@@ -50,37 +75,44 @@ int test_constructor()
 
    for (int i = 0; i < Trials; ++i)
    {
-      NumericType n = uniform_integral_number<NumericType>();
+      NumericType n = get_rand<NumericType>();
 
       double_float_t d(n);
 
-      double_float_t::rep_type rep(d.rep());
+      typename double_float_t::rep_type rep(d.rep());
       double_float_t::normalize_pair(rep);
 
       // Check if representation of the cpp_double_float is not normalized
       if (rep != d.rep())
       {
-         std::cerr << "[FAILED]\n" << typeid(NumericType).name() << " = " << n
-           << " (cpp_double_float<" << typeid(FloatingPointType).name() << "> = " << d.get_raw_str() << ")" << std::endl;
-         return -1;
+         std::cerr << "[FAILED]\nabnormal representation for " << typeid(NumericType).name() << " = " << n
+                   << " (cpp_double_float<" << typeid(FloatingPointType).name() << "> = " << d.get_raw_str() << ")" << std::endl;
+
+         result_is_ok = false;
+
+         break;
       }
 
-      // Check if value is accurately represented and correctly rounded
+      NumericType n_prime(n);
+
+      // Check if value is accurately represented
       constexpr int NumericTypeBits      = std::numeric_limits<NumericType>::digits;
-      constexpr int MaxRepresentableBits = std::numeric_limits<FloatingPointType>::digits * 2;
-      const NumericType RoundedBitsMask      = (NumericType(1) << (NumericTypeBits - MaxRepresentableBits)) - 1;
-      const NumericType RoundedMargin        = (RoundedBitsMask + 1) >> 1;
+      constexpr int MaxRepresentableBits = std::numeric_limits<double_float_t>::digits;
 
-      NumericType   n_prime(n);
+      // Round correctly if the integral type has more precision than what can be represented
+      if (NumericTypeBits > MaxRepresentableBits)
+      {
+         const NumericType RoundedBitsMask = (NumericType(1) << (NumericTypeBits - MaxRepresentableBits)) - 1;
+         const NumericType RoundedMargin   = (RoundedBitsMask + 1) >> 1;
 
-      n_prime >>= NumericTypeBits - MaxRepresentableBits;
-      // Round correctly
-      if (n & RoundedBitsMask > RoundedMargin)
-         n_prime |= 1;
-      if (n & RoundedBitsMask < RoundedMargin)
-         n_prime &= ~NumericType(1);
+         n_prime >>= NumericTypeBits - MaxRepresentableBits;
+         if ((n & RoundedBitsMask) > RoundedMargin)
+            n_prime |= 1;
+         else if ((n & RoundedBitsMask) < RoundedMargin)
+            n_prime &= ~NumericType(1);
 
-      n_prime <<= NumericTypeBits - MaxRepresentableBits;
+         n_prime <<= NumericTypeBits - MaxRepresentableBits;
+      }
 
       if (std::abs(signed(n_prime - n)) < std::abs(signed(static_cast<NumericType>(d) - n)))
       {
@@ -88,17 +120,77 @@ int test_constructor()
                    << "> = 0x" << std::hex << static_cast<NumericType>(d)
                    << " (expected 0x" << std::hex << n_prime << ")"
                    << std::endl;
-         return -1;
+
+         result_is_ok = false;
+
+         break;
       }
    }
 
-   std::cout << "[PASSED] (" << Trials << " cases tested)" << std::endl;
-   return 0;
+   std::cout << "ok (" << Trials << " cases tested)" << std::endl;
+
+   return result_is_ok;
+}
+
+template <typename FloatingPointType, typename OtherFloatType, typename std::enable_if<!std::is_integral<OtherFloatType>::value>::type const* = nullptr>
+bool test_constructor()
+{
+   bool result_is_ok = true;
+
+   constexpr int Trials = 50000;
+
+   std::string   type_name = typeid(OtherFloatType).name();
+   size_t        idx;
+   if ((idx = type_name.rfind(":")) != std::string::npos)
+      type_name = type_name.substr(idx + 1, type_name.size());
+
+   std::cout << "Testing constructor for ";
+   std::cout.width(30);
+   std::cout << type_name << "... ";
+   using double_float_t = boost::multiprecision::backends::cpp_double_float<FloatingPointType>;
+
+   for (int i = 0; i < Trials; ++i)
+   {
+      OtherFloatType n = get_rand<OtherFloatType>();
+
+      double_float_t d(n);
+
+      typename double_float_t::rep_type rep(d.rep());
+      double_float_t::normalize_pair(rep);
+
+      // Check if representation of the cpp_double_float is not normalized
+      if (rep != d.rep())
+      {
+         std::cerr << "[FAILED]\nabnormal representation for " << typeid(OtherFloatType).name() << " = " << n
+                   << " (cpp_double_float<" << typeid(FloatingPointType).name() << "> = " << d.get_raw_str() << ")" << std::endl;
+
+         result_is_ok = false;
+
+         break;
+      }
+
+      // Check if the binary digits match (work in progress...)
+      /*const int      DigitsToMatch  = std::min(std::numeric_limits<double_float_t>::digits, std::numeric_limits<OtherFloatType>::digits);
+      int digits_matched = 0;
+
+      OtherFloatType n_prime(n);
+      double_float_t d_prime(d);
+
+      while (n_prime > 1)
+      {
+         n_prime /= 2;
+         d_prime /= 2;
+      }*/
+   }
+
+   std::cout << "ok (" << Trials << " cases tested)" << std::endl;
+
+   return result_is_ok;
 }
 
 // Test compilation, constructors, basic operatory
 template <typename FloatingPointType>
-int test_constructors()
+bool test_constructors()
 {
    using double_float_t = boost::multiprecision::backends::cpp_double_float<FloatingPointType>;
    double_float_t a, b;
@@ -106,32 +198,39 @@ int test_constructors()
    std::cout << "Testing cpp_double_float< " << typeid(FloatingPointType).name() << " >...\n==="
              << std::endl;
 
-   int e = 0;
-   e += test_constructor<FloatingPointType, long long int>();
-   e += test_constructor<FloatingPointType, unsigned long long int>();
-   e += test_constructor<FloatingPointType, long int>();
-   e += test_constructor<FloatingPointType, unsigned long int>();
-   e += test_constructor<FloatingPointType, short int>();
-   e += test_constructor<FloatingPointType, unsigned short int>();
-   e += test_constructor<FloatingPointType, signed char>();
-   e += test_constructor<FloatingPointType, unsigned char>();
+   bool result_is_ok = true;
 
-   if (e == 0)
-      std::cout << "PASSED ALL TESTS";
+   result_is_ok &= test_constructor<FloatingPointType, long long int>();
+   result_is_ok &= test_constructor<FloatingPointType, unsigned long long int>();
+   result_is_ok &= test_constructor<FloatingPointType, long int>();
+   result_is_ok &= test_constructor<FloatingPointType, unsigned long int>();
+   result_is_ok &= test_constructor<FloatingPointType, short int>();
+   result_is_ok &= test_constructor<FloatingPointType, unsigned short int>();
+   result_is_ok &= test_constructor<FloatingPointType, signed char>();
+   result_is_ok &= test_constructor<FloatingPointType, unsigned char>();
+   result_is_ok &= test_constructor<FloatingPointType, float>();
+   result_is_ok &= test_constructor<FloatingPointType, double>();
+   result_is_ok &= test_constructor<FloatingPointType, float>();
+   result_is_ok &= test_constructor<FloatingPointType, boost::multiprecision::backends::cpp_double_float<float>>();
+   result_is_ok &= test_constructor<FloatingPointType, boost::multiprecision::backends::cpp_double_float<double>>();
+
+   if (result_is_ok)
+      std::cout << "PASSED all tests";
    else
-      std::cout << "FAILED " << -e << " TESTS";
-   std::cout << std::endl << std::endl;
+      std::cout << "FAILED some test(s)";
+   std::cout << std::endl
+             << std::endl;
 
-   return e == 0 ? 0 : -1;
+   return result_is_ok;
 }
 } // namespace test_cpp_double_constructors
 
 int main()
 {
-   test_cpp_double_constructors::test_constructors<float>();
-   test_cpp_double_constructors::test_constructors<double>();
+   const bool result_float_is_ok  = test_cpp_double_constructors::test_constructors<float>();
+   const bool result_double_is_ok = test_cpp_double_constructors::test_constructors<double>();
 
-   std::cin.get();
+   const bool result_is_ok = (result_float_is_ok && result_double_is_ok);
 
-   return 0;
+   return (result_is_ok ? 0 : -1);
 }
